v0.4.2 Clearchat Confirm

+++ commands/utility/clearchat.js
@@ -1,23 +1,32 @@
+const usedChannels = [];
+
 async function clearChat(channel) {
-    console.log('Clearing all chat from ' + channel.name);
-    let fetched;
-    let error = false;
-    do {
-        try {
-            fetched = await channel.messages.fetch({ limit: 100 });
-            await channel.bulkDelete(fetched);
+    if (usedChannels.includes(channel)) {
+        channel.cmdreply.send('กำลังลบข้อความทั้งหมด', { 'flags': 64 });
+        console.log('Clearing all chat from ' + channel.name);
+        let fetched;
+        let error = false;
+        do {
+            try {
+                fetched = await channel.messages.fetch({ limit: 100 });
+                await channel.bulkDelete(fetched);
+            }
+            catch (err) {
+                console.error(err.message);
+                channel.send(err.message);
+                error = true;
+            }
+            finally {
+                channel.bulkDelete(fetched, true);
+            }
         }
-        catch (err) {
-            console.error(err.message);
-            channel.send(err.message);
-            error = true;
-        }
-        finally {
-            channel.bulkDelete(fetched, true);
-        }
+        while (fetched.size >= 2 && !error);
+        console.log('Chat cleared.');
     }
-    while (fetched.size >= 2 && !error);
-    console.log('Chat cleared.');
+    else {
+        channel.cmdreply.send(':warning: ยังไม่เคยใช้คำสั่งนี้ในช่องนี้มาก่อนในช่วงเร็ว ๆ นี้\n:exclamation: พิมพ์คำสั่งอีกครั้งเพื่อยืนยันการใช้งาน');
+        usedChannels.push(channel);
+    }
 }
 
 module.exports = {
@@ -26,12 +35,11 @@ module.exports = {
     description: 'ลบแชทใน channel ทั้งหมด (Admin เท่านั้น)',
     permissions: 'ADMINISTRATOR',
     guildOnly: true,
-    cooldown: 15,
+    cooldown: 5,
     execute(message, args) {
         clearChat(message.channel);
     },
     executeSlash(interaction, args) {
-        interaction.channel.cmdreply.send('กำลังลบข้อความทั้งหมด', { 'flags': 64 });
         clearChat(interaction.channel);
     }
 };
